Rename Team to Transport
Fixes #50
