+++ src/main/webapp/WEB-INF/tags/beam-transport-page.tag
@@ -1,4 +1,4 @@
-<%@tag description="The Beam Team Page Template Tag" pageEncoding="UTF-8"%>
+<%@tag description="The Beam Transport Page Template Tag" pageEncoding="UTF-8"%>
 <%@taglib prefix="fn" uri="http://java.sun.com/jsp/jstl/functions"%>
 <%@taglib prefix="s" uri="http://jlab.org/jsp/smoothness"%>
 <%@taglib prefix="t" tagdir="/WEB-INF/tags"%>
@@ -15,9 +15,9 @@
     </jsp:attribute>
     <jsp:attribute name="secondaryNavigation">
         <ul>
-            <li${'/beam-team/weekly-tune' eq currentPath ? ' class="current-secondary"' : ''}><a href="${pageContext.request.contextPath}/beam-team/weekly-tune">Weekly Tune</a></li>
-            <li${'/beam-team/tune-incidents' eq currentPath ? ' class="current-secondary"' : ''}><a href="${pageContext.request.contextPath}/beam-team/tune-incidents">Tune Incidents</a></li>
-            <li${'/beam-team/tune-comparison' eq currentPath ? ' class="current-secondary"' : ''}><a href="${pageContext.request.contextPath}/beam-team/tune-comparison">Tune Comparison</a></li>
+            <li${'/beam-transport/weekly-tune' eq currentPath ? ' class="current-secondary"' : ''}><a href="${pageContext.request.contextPath}/beam-transport/weekly-tune">Weekly Tune</a></li>
+            <li${'/beam-transport/tune-incidents' eq currentPath ? ' class="current-secondary"' : ''}><a href="${pageContext.request.contextPath}/beam-transport/tune-incidents">Tune Incidents</a></li>
+            <li${'/beam-transport/tune-comparison' eq currentPath ? ' class="current-secondary"' : ''}><a href="${pageContext.request.contextPath}/beam-transport/tune-comparison">Tune Comparison</a></li>
         </ul>
     </jsp:attribute>
     <jsp:body>
