+++ src/main/webapp/WEB-INF/views/beam-transport/tune-comparison.jsp
@@ -7,7 +7,7 @@
 <%@taglib prefix="t" tagdir="/WEB-INF/tags"%> 
 <jsp:useBean id="now" class="java.util.Date"/>
 <c:set var="title" value="Tune Comparison"/>
-<t:beam-team-page title="${title}">  
+<t:beam-transport-page title="${title}">
     <jsp:attribute name="stylesheets">
         <link rel="stylesheet" type="text/css" href="${pageContext.request.contextPath}/resources/v${initParam.releaseNumber}/css/downtime-report.css"/>
     </jsp:attribute>
@@ -180,4 +180,4 @@
             <button id="exit-fullscreen-button">Exit Full Screen</button>
         </div>
     </jsp:body>         
-</t:beam-team-page>
+</t:beam-transport-page>
